Application routing implementation
Added the app router.
Added the new app routing module.
Added the team and the teams page routes.
Documented the new classes and methods.
Modified the template to integrate the router outlet tags.
refactored the team component to stop the use of @Input variable.
Injected the router and the activated route where necessary.
Updated the app module file.